connected mongodb. saved a product into db by using insomnia

+++ src/main/java/be/fooda/backend/basket/service/flow/ProductFlow.java
@@ -2,11 +2,15 @@ package be.fooda.backend.basket.service.flow;
 
 import be.fooda.backend.basket.dao.ProductRepository;
 import be.fooda.backend.basket.model.dto.ProductResponse;
+import be.fooda.backend.basket.model.http.HttpFailureMessages;
 import be.fooda.backend.basket.service.exception.ResourceNotFoundException;
 import be.fooda.backend.basket.service.mapper.ProductMapper;
 import be.fooda.backend.basket.model.entity.ProductEntity;
 import be.fooda.backend.basket.model.dto.CreateProductRequest;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,29 +25,55 @@ public class ProductFlow {
     private final ProductMapper productMapper;
 
 
-    public Optional<ProductEntity> findById(String id){
-        return productRepository.findById(id);
+    public void createProduct(CreateProductRequest request){
+
+        // IF(NULL)
+        if(Objects.isNull(request)){
+            // THROW_EXCEPTION
+            throw new NullPointerException(HttpFailureMessages.FAILED_TO_CREATE_PRODUCT.getDescription());
+        }
+
+       /* boolean exists = productRepository.existsByNameAndStore_EStoreId(request.getName(), request.getStore().getEStoreId());
+
+        //  IF(PRODUCT_EXISTS)
+        if(exists){
+            throw new ResourceNotFoundException(HttpFailureMessages.PRODUCT_ALREADY_EXISTS.getDescription());
+        }*/
+
+        // MAP_DTO_TO_ENTITY
+        ProductEntity entity = productMapper.toEntity(request);
+
+        // SAVE_TO_DB(ENTITY)
+        productRepository.save(entity);
+
+        //return productMapper.toResponse(entity);
+    }
+
+    public List<ProductResponse> findAll(int pageNo, int pageSize){
+
+        Pageable pageable = PageRequest.of(pageNo-1,pageSize);
+
+        Page<ProductEntity> pages = productRepository.findAll(pageable);
+
+        return productMapper.toResponses(pages.toList());
     }
 
     public List<ProductEntity> findAll(){
+
         return productRepository.findAll();
     }
 
-    public ProductResponse createProduct(CreateProductRequest request){
+    public ProductResponse findById(String id){
 
-        if(Objects.isNull(request)){
-            throw new NullPointerException("There is no request in the body");
-        }
+        ProductEntity productEntity = productRepository.findById(id)
+                .orElseThrow(()-> new ResourceNotFoundException(HttpFailureMessages.PRODUCT_DOES_NOT_EXIST_IN_BASKET.getDescription()));
 
-        boolean exists = productRepository.existsByNameAndStore_EStoreId(request.getName(), request.getStore().getEStoreId());
-
-        if(exists){
-            throw new ResourceNotFoundException("Product NOT found");
-        }
-        ProductEntity entity = productRepository.save(productMapper.toEntity(request));
-
-        return productMapper.toResponse(entity);
+        return productMapper.toResponse(productEntity);
     }
+
+
+
+
     public List<ProductEntity> findByUser(String eUserId){
         return productRepository.findByUser(eUserId);
     }
@@ -69,6 +99,9 @@ public class ProductFlow {
     }
 
     public void deleteById(String id){
+        Optional<ProductEntity> product = productRepository.findById(id);
+        if (!product.isPresent())
+            throw new ResourceNotFoundException("");
         productRepository.deleteById(id);
     }
